rename and update examples/paralle.py
